+++ test/typings/either-spec.ts
@@ -65,7 +65,6 @@ const messageCopy: string = Left<number, string>(404)
     .ap(unpacked.map(m => (p: string) => p + ' ' + m))
     .cata(e => e, v => v);
 
-const contains: boolean = wrappedGreeting.contains("test");
 Right<number, string>("hello").forEach((str:string) => console.log(str));
 Left<string, number>("none").forEachLeft((str:string) => console.log("oops"));